FastAPI/requirements.txt
anyio==3.6.2
asttokens==2.2.1
attrs==23.2.0
autopep8==2.0.2
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
executing==1.2.0
fastapi==0.95.1
filelock==3.12.0
Flask==2.3.2
fonttools==4.39.3
greenlet==3.0.3
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
orjson==3.8.12
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
PyMySQL==1.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
requests==2.30.0
rsa==4.9
seaborn==0.12.2
selenium==4.18.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.30
stack-data==0.6.2
starlette==0.26.1
tornado==6.3.1
traitlets==5.9.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2023.3
ujson==5.7.0
urllib3==2.0.2
uvicorn==0.22.0
virtualenv==20.23.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.3
wsproto==1.2.0
